+++ src/compile.rs
@@ -23,7 +23,9 @@ pub fn compile(
     // TODO: What do we do when there are multiple bin targets?
     match bindings_crate {
         BridgeModel::Bin => shared_args.push("--bins"),
-        BridgeModel::Cffi | BridgeModel::Bindings(_) => shared_args.push("--lib"),
+        BridgeModel::Cffi | BridgeModel::Bindings(_) | BridgeModel::BindingsAbi3 => {
+            shared_args.push("--lib")
+        }
     }
 
     shared_args.extend(context.cargo_extra_args.iter().map(String::as_str));
@@ -40,8 +42,9 @@ pub fn compile(
         .map(String::as_str)
         .collect();
 
+    // https://github.com/PyO3/pyo3/issues/88#issuecomment-337744403
     if context.target.is_macos() {
-        if let BridgeModel::Bindings(_) = bindings_crate {
+        if let BridgeModel::Bindings(_) | BridgeModel::BindingsAbi3 = bindings_crate {
             let mac_args = &["-C", "link-arg=-undefined", "-C", "link-arg=dynamic_lookup"];
             rustc_args.extend(mac_args);
         }