+++ Changelog.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) (for the cli, not for the crate).
 
+## Unreleased
+
+### Added
+
+ * Added support for building abi3 wheels with pyo3 0.13
+
 ## 0.8.3 - 2020-08-17
 
 ### Added